Pushed v1.37 code
- The 4-tap cubic spline resampling interpolation has now been replaced with 8-tap windowed-sinc (polyphase) for better audio quality
- The mixer is now using 64-bit float precision instead of 32-bit float precision. We only mix up to 32 channels anyway, which is not very CPU intensive
- Volume/panning envelope interpolation is now using floating point precision instead of 8.8 fixed-point precision
- Fixed a bug where the scopes would not shut down if you triggered an instrument that is not allocated (empty)
- Small code cleanup

+++ src/ft2_edit.c
@@ -963,7 +963,7 @@ void doTranspose(void)
 	countOverflowingNotes(lastInsMode, lastTranspMode, lastTranspVal);
 	if (transpDelNotes > 0)
 	{
-		sprintf(text, "%d note(s) will be erased! Proceed?", transpDelNotes);
+		sprintf(text, "%d note(s) will be erased! Proceed?", (int32_t)transpDelNotes);
 		if (okBox(2, "System request", text) != 1)
 			return;
 	}
